requirements.txt
argh==0.26.2
Brotli==1.0.4
certifi==2018.8.13
chardet==3.0.4
cloudinary==1.12.0
dj-database-url==0.5.0
Django==2.1
django-debug-toolbar==1.9.1
django-dotenv==1.4.2
django-ipware==2.1.0
django-redis==4.9.0
django-sslserver==0.20
djangorestframework==3.8.2
funcy==1.10.3
gunicorn==19.9.0
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
mock==2.0.0
msgpack==0.5.6
pathtools==0.1.2
pbr==4.2.0
ply==3.11
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
PyYAML==3.13
redis==2.10.6
six==1.11.0
slimit==0.8.1
sqlparse==0.2.4
urllib3==1.23
watchdog==0.8.3
whitenoise==4.0
